Complete the Collecting Signatures problem
The key is to sort the segments from smallest right point to largest right point. After soring, for any segments i and j such that segment j is to the right of segment i, the segment j either spans over the right point of segment i or does not touch segment i at all. The solution is the first smallest right point, then the next right point of an UNCOVERED segment, and so on...