Complete Maximum Number of Prizes problem
The key is to start with 1 and work the way up. Don't forget all the summands must be distinct from each other. Subtract the remaining sum until zero or until reach the number where it cannot be expressed by two summands (due to repeating terms). If in the case of repeating terms, take final unexpressable sum as the final summand.